+++ js/index.js
@@ -1,10 +1,7 @@
-import 'url-pattern';
-import 'custom-elements';
-import 'highlight.js';
 import '@shgysk8zer0/polyfills';
 import '@shgysk8zer0/kazoo/dom.js';
 import '@kernvalley/components/ad.js';
 import '@kernvalley/components/events.js';
 import '@shgysk8zer0/components/github/user.js';
-import 'leaflet';
-import 'firebase/firebase-firestore.js';
+
+console.log(import.meta.url);