+++ postcss.config.js
@@ -1,4 +1,4 @@
 /* eslint-env node */
-import { getConfig } from './postcss/utils.js';
+import { getConfig } from '@shgysk8zer0/css-utils/postcss';
 
 export default getConfig({ inline: false });